moved getSignature and pipe into extractSignature as private functions
