+++ example.js
@@ -4,7 +4,8 @@ const hmac = require('.')
 
 module.exports = function (fastify, options, next) {
   fastify.register(hmac, {
-    sharedSecret: 'topSecret'
+    sharedSecret: 'topSecret',
+    extractSecret: () => {}
   })
 
   fastify.post('/', (req, reply) => {
