+++ lib/extractSignature.js
@@ -1,11 +1,24 @@
 'use strict'
 
-const pipe = require('./pipe')
 const parseSignatureHeader = require('./parseSignatureHeader')
-const getSignature = require('./getSignature')
 
+// Public Functions
 function extractSignature (req) {
   return pipe(parseSignatureHeader, getSignature)(req)
+}
+
+// Private Functions
+const pipe = (...functions) => (value) =>
+  functions.reduce(
+    (currentValue, currentFunction) => currentFunction(currentValue),
+    value
+  )
+
+const getSignature = ({ signature }) => {
+  if (!signature) {
+    throw new Error('Could not get signature: no signature found.')
+  }
+  return signature
 }
 
 module.exports = extractSignature
