+++ lib/extractSignature.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { test } = require('tap')
-const getSignature = require('./getSignature')
+const extractSignature = require('./extractSignature')
 
 test('Extracts a signature from a valid signature header', async ({
   equal
@@ -9,18 +9,13 @@ test('Extracts a signature from a valid signature header', async ({
   const tests = [
     {
       description: 'Signature extracted from valid signature header',
-      input: { signature: '1d3g5h7' },
+      input: { headers: { signature: 'signature=1d3g5h7, biff=buzz' } },
       expected: '1d3g5h7'
-    },
-    {
-      description: "Dictionary respects '=' in field values as valid character",
-      input: { signature: '2d3g=5h7' },
-      expected: '2d3g=5h7'
     }
   ]
 
   tests.forEach(({ description, input, expected }) => {
-    equal(getSignature(input), expected, description)
+    equal(extractSignature(input), expected, description)
   })
 })
 
@@ -30,7 +25,7 @@ test('Throws an error when signature cannot be extracted', async ({
   const tests = [
     {
       description: 'No signature field in signature header returns an error',
-      input: { notSignature: 'buzz' },
+      input: { headers: { signature: 'biff=buzz' } },
       expected: new Error('Could not get signature: no signature found.')
     }
   ]
@@ -38,7 +33,7 @@ test('Throws an error when signature cannot be extracted', async ({
   tests.forEach(({ description, input, expected }) => {
     throws(
       function () {
-        getSignature(input)
+        extractSignature(input)
       },
       expected,
       description
