--- lib/pipe.js
@@ -1,9 +1,0 @@
-'use strict'
-
-const pipe = (...functions) => (value) =>
-  functions.reduce(
-    (currentValue, currentFunction) => currentFunction(currentValue),
-    value
-  )
-
-module.exports = pipe
